+++ dzVents/documentation/history.md
@@ -4,6 +4,7 @@
 - Add setHue, setColor, setHex, getColor for RGBW(W) devices
 - Add setDescription for devices, groups and scenes
 - Add volumeUp / volumeDown for Logitech Media Server (LMS)
+- Changed domoticz.utils.fromJSON (add optional fallback param) 
 
 [2.4.15] 
 - Add option to use camera name in snapshot command